refactor: only run systemctl when systemd is running as init
Based on reviewer feedback, changed the logic to only attempt
systemctl commands when systemd is actually running as PID 1
(has_systemd() returns true).

This is cleaner and more explicit:
- Systems with systemd running: use systemctl
- Systems with service command: use service
- Container environments: skip service management gracefully

The previous approach of running systemctl even when systemd
wasn't running was for image portability (container->VM), but
this is a niche use case.

Signed-off-by: Gajesh Bhat <gajeshbht@gmail.com>